convert.py: Extract alignments from frames if they don't exist
